initial commit for handling bad sstables
I am not at all happy about this change.  The basic idea is to work
around bad rows by keeping track of where we are in the index files.
The problem is basically that this mechanism is far reaching and hard to
test.  Possibly before this can be fully committed we need to bite the
bullet and create unit tests for the Hadoop features of the code
(inputformats, splits, recordreaders).